Disable ULP timer on wakeup
When I removed the bit disabling the wakeup timer for the ULP processor
it looks like it wound up causing a race during initialization between
the main processor seeing the button bits and the ULP processor
clearing them. So we need to make sure to turn off the ULP processor.